custom-elements/instantiating-custom-elements/extensions-to-document-interface/created-callback-test-001.html

<!DOCTYPE html>
<html>
<head>
<title>Created callback should be called if createElement is called</title>
<meta name="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<meta name="assert" content="Created callback should be called if createElement is called">
<link rel="help" href="http://www.w3.org/TR/custom-elements/#types-of-callbacks">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src='../../testcommon.js'></script>
<link rel="stylesheet" href="/resources/testharness.css">
</head>
<body>
<div id="log"></div>
<script type="text/javascript">

test(function() {
    var doc = newHTMLDocument();
    
    var callbackCalled = false;
    
    var proto = Object.create(HTMLElement.prototype);
    proto.createdCallback = function () {
        callbackCalled = true;
    }
    
    var name = 'x-a-1';
    
    doc.register(name, {prototype: proto});
    
    var customElement = doc.createElement(name);

    assert_true(callbackCalled, 'Callback created was not called');
    
}, 'Created callback should be called if createElement is called. Type is not specified');


test(function() {
    var doc = newHTMLDocument();
    
    var callbackCalled = false;
    
    var proto = Object.create(HTMLElement.prototype);
    proto.createdCallback = function () {
        callbackCalled = true;
    }
    
    var name = 'x-a-2';
    
    doc.register(name, {prototype: proto});
    
    var customElement = doc.createElement(name, name);

    assert_true(callbackCalled, 'Callback created was not called');
    
}, 'Created callback should be called if createElement is called. Type is specified');


test(function() {
    var doc = newHTMLDocument();
    
    var callbackCalled = false;
    
    var proto = Object.create(HTMLElement.prototype);
    proto.createdCallback = function () {
        callbackCalled = true;
    }
    
    var name = 'x-a-3';
    
    var customElement = doc.createElement(name);

    assert_false(callbackCalled, 'Callback created was not called');
    
    doc.register(name, {prototype: proto});
    
    assert_true(callbackCalled, 'Callback created was not called');
    
}, 'Created callback should be called if createElement is called. ' +
    'Element is unresolved at first');


test(function () {
    for (var i = 0; i < HTML5_ELEMENTS.length; i++) {
        var element = HTML5_ELEMENTS[i];
        
        var doc = newHTMLDocument();
        
        var obj = doc.createElement(element);
        
        var name = 'x-' + element;
        
        var proto = Object.create(obj.constructor.prototype);
        proto.createdCallback = function () {
            callbackCalled = true;
        }    
        
        doc.register(name, {prototype: proto, extends: element});
        
        var createElement = doc.createElement(element, name);
            
        assert_true(callbackCalled, 'Callback created was not called');
    }
    
}, 'Created callback should be called if createElement is called. Test different types');

</script>
</body>
</html>
